onors/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
boto3==1.7.76
botocore==1.10.76
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3
defusedxml==0.5.0
Django==2.1
django-allauth==0.36.0
django-cors-headers==2.4.0
django-encrypted-model-fields==0.5.8
django-rest-auth==0.9.3
django-storages==1.6.6
django-stringfield==0.3.0
djangorestframework==3.8.2
docutils==0.14
envs==1.2.6
idna==2.7
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
mysqlclient==1.3.13
oauthlib==2.1.0
Pillow==5.2.0
pycparser==2.18
pycrypto==2.6.1
python-dateutil==2.7.3
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
simple-crypt==4.1.7
six==1.11.0
terminaltables==3.1.0
urllib3==1.23
